+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog fastify-plugin
 
 ## Upcoming changes
+- [FIX] [FLR] replace dep keycloak-connect by keycloak-connect-hoppen
 
 ## Version 0.3.0 - 2022-08-10
 - [FEA] [FLR] upgrade to fastify-plugin 4.1.0